Adding a new InlineAgent ChatModel and refactoring agents codebase
