+++ libs/aws/langchain_aws/agents/__init__.py
@@ -1,8 +1,15 @@
 from langchain_aws.agents.base import (
-    BedrockAgentAction,
-    BedrockAgentFinish,
     BedrockAgentsRunnable,
     BedrockInlineAgentsRunnable,
 )
 
-__all__ = ["BedrockAgentAction", "BedrockAgentFinish", "BedrockAgentsRunnable", "BedrockInlineAgentsRunnable"]
+from langchain_aws.agents.types import (
+    BedrockAgentAction,
+    BedrockAgentFinish,
+)
+
+from langchain_aws.agents.inline_chat import (
+    BedrockInlineAgentsChatModel,
+)
+
+__all__ = ["BedrockAgentAction", "BedrockAgentFinish", "BedrockAgentsRunnable", "BedrockInlineAgentsRunnable", "BedrockInlineAgentsChatModel"]
